backend/requirements.txt
fastapi
uvicorn
anthropic
groq
networkx
celery
redis
pydantic
httpx
python-dotenv
loguru
flask
requests
google-auth
passlib
bcrypt